+++ CHANGELOG.md
@@ -16,18 +16,19 @@
 - **client-logger:** update client logger
 - **discord:** better embed structures
 - **discord:** added multipart uploads
+- **entry:** added json marshaler support
 - **error_node:** indent now set to 3 spaces
 - **http-client:** wip update
 - **implError:** exported implError as ErrorNode
-- **query:** added CollectErrors query
 - **query:** added bottom error query
+- **query:** added CollectErrors query
 - **tower:** added code block marshaler pattern
 - **tower-discord:** implemented file upload native discord
 - **tower-discord:** added thread id in metadata
-- **tower-http:** added logging middleware to respond error
 - **tower-http:** added client logger
 - **tower-http:** added respond body logger
 - **tower-http:** added logging middleware
+- **tower-http:** added logging middleware to respond error
 
 ### Miscellaneous
 - synchronize go mod with new tag on tower